THATZ IMPOSSIBLE
Hmm
I remember years ago
Someone told me I should take
Caution when it comes to love, I did
And you were strong and I was not
My illusion, my mistake
I was careless, I forgot, I did
And now
When all is done, there is nothing to say
You have gone and so effortlessly
You have won, you can go ahead tell them
Tell them all I know now
Shout it from the roof tops
Write it on the sky line
All we had is gone now
Tell them I was happy
And my heart is broken
All my scars are open
Tell them what I hoped would be impossible,
Impossible
Impossible
Impossible
Falling out of love is hard
Falling for betrayal is worse
Broken trust and broken hearts
I know, I know
And thinking all you need is there
Building faith on love and words
Empty promises will wear
I know
I know and now
When all is done, there is nothing to say
And if you're done with embarrassing me
On your own you can go ahead tell them
Tell them all I know now
Shout it from the roof tops
Write it on the sky line
All we had is gone now
Tell them I was happy
And my heart is broken
All my scars are open
Tell them what I hoped would be impossible,
Impossible
Impossible
Impossible
I remember years ago
Someone told me I should take
Caution when it comes to love
I did
Tell them all I know now
Shout it from the roof tops
Write it on the sky line
All we had is gone now
Tell them I was happy
And my heart is broken
Oh what I hoped would be impossible,
Impossible (impossible)
Impossible (impossible)
(Impossible)
impossible
(Impossible) impossible
(Impossible)
Impossible

+++ src/servlet/DeleteActivity.java
@@ -33,10 +33,16 @@ public class DeleteActivity extends HttpServlet {
 		HttpSession s = request.getSession(true);
 		ActivityDB actdb = new ActivityDB();
 		
-	
+		String check = request.getParameter("action");
 		String actId =(request.getParameter("actId"));
+		if (check.equalsIgnoreCase("Invalid")) {
 		actdb.deleteActivity(actId);
-		response.sendRedirect("ActList");
+		response.getWriter().append("N");}
+		else {
+			actdb.restoreActivity(actId);
+			response.getWriter().append("Y");
+		}
+		
 	}
 
 	/**
